+++ tests/api/BaseApiTestCase.php
@@ -1,5 +1,5 @@
 <?php
-namespace App\Api\Test;
+namespace App\Tests\api;
 
 use ApiPlatform\Core\Bridge\Symfony\Bundle\Test\ApiTestCase;
 use ApiPlatform\Core\Bridge\Symfony\Bundle\Test\Client;
